Minimal postprocessing and GL-related refactor

+++ src/graphics/Viewport.h
@@ -14,8 +14,8 @@ public:
     virtual uint getWidth() const;
     virtual uint getHeight() const;
 
-    glm::vec2 size() const {
-        return glm::vec2(width, height);
+    glm::ivec2 size() const {
+        return glm::ivec2(width, height);
     }
 };
